chore(commands): code cleanup of some commands
